backend/requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.34.0,<1.0
httpx>=0.28.0,<1.0
pydantic-settings>=2.7.0,<3.0
python-dotenv>=1.0.0,<2.0
sqlalchemy[asyncio]>=2.0.36,<3.0
asyncpg>=0.30.0,<1.0
alembic>=1.14.0,<2.0
PyJWT>=2.10.0,<3.0
cryptography>=44.0.0,<45.0
